# docs/openapi.yaml
openapi: 3.1.0
info:
  title: LegalHub API
  description: |
    API REST do LegalHub (gestão jurídica). Autenticação via JWT Bearer, exceto login.

    Fonte de contratos: schemas Zod em `backend/src/modules/**/*.schemas.ts`.
  version: 0.1.0
  contact:
    name: LegalHub
servers:
  - url: http://localhost:3001
    description: Desenvolvimento local
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Sessão criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/session:
    get:
      tags: [Auth]
      summary: Validar sessão atual
      operationId: getSession
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Sessão válida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Encerrar sessão
      operationId: logout
      security:
        - bearerAuth: []
      description: Revoga a sessão no servidor. Token opcional; retorna 204 mesmo sem token válido.
      responses:
        "204":
          description: Sessão encerrada
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token retornado por POST /api/auth/login
...
  responses:
...
    Unauthorized:
      description: Não autorizado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: Recurso não encontrado
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          description: Papel do usuário (ex. admin, member)
    SessionResponse:
      type: object
      required: [token, expiresAt, user]
      properties:
        token:
          type: string
        expiresAt:
          type: string
          format: date-time
          description: Data de expiração da sessão (ISO 8601)
        user:
          $ref: "#/components/schemas/User"
...